+++ libelement/src/object_model/object_model_builder.hpp
@@ -17,7 +17,7 @@ namespace element
     {
         const auto& file_info = context->src_context->file_info.at(ast->nearest_token->file_name);
         const std::string* filename = file_info.file_name.get();
-        const std::string* line_in_source = file_info.source_lines[ast->nearest_token->line - 1].get();
+        const std::string* line_in_source = file_info.source_lines[static_cast<std::size_t>(ast->nearest_token->line) - 1].get();
         t->source_info = source_information(
             ast->nearest_token->line,
             ast->nearest_token->character,
